rename source tab to credit
