+++ src/components/organisms/SidePanel/ManuscriptSidePanel/Tabs/CreditTab/ContributorsSection.tsx
