+++ src/components/organisms/SidePanel/ManuscriptSidePanel/Tabs/CreditTab/Credits.tsx
