+++ src/components/organisms/SidePanel/ManuscriptSidePanel/Tabs/CreditTab/CreditsPreview.tsx
