+++ src/components/organisms/SidePanel/ManuscriptSidePanel/Tabs/CreditTab/OrganizationsSection.tsx
