+++ src/components/organisms/SidePanel/ManuscriptSidePanel/Tabs/CreditTab/PropertyTagsSection.tsx
