+++ src/components/organisms/SidePanel/ManuscriptSidePanel/index.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import { FlexColumn, FlexRowSpaceBetween } from "@components/styled";
 import SidePanel from "@components/organisms/SidePanel";
 import HistoryTab from "@components/organisms/SidePanel/ManuscriptSidePanel/Tabs/History/HistoryTab";
-import SourceTab from "@components/organisms/SidePanel/ManuscriptSidePanel/Tabs/SourceTab";
+import CreditTab from "@src/components/organisms/SidePanel/ManuscriptSidePanel/Tabs/CreditTab";
 import PerfectScrollbar from "react-perfect-scrollbar";
 import PanelCloseButton from "@components/atoms/PanelCloseButton";
 import ManuscriptAttributesSection from "@src/components/organisms/ManuscriptAttributesSection";
@@ -292,11 +292,11 @@ const ManuscriptSidePanel = (props: ManuscriptSidePanelProps) => {
               </p>
             </SwitchButton>
             <SwitchButton
-              isSelected={researchPanelTab === ResearchTabs.source}
-              onClick={() => onSetResearchPanelTab(ResearchTabs.source)}
+              isSelected={researchPanelTab === ResearchTabs.credit}
+              onClick={() => onSetResearchPanelTab(ResearchTabs.credit)}
             >
               <p className="text-xs flex justify-center items-center h-full">
-                Source
+                Credit
               </p>
             </SwitchButton>
           </SwitchBar>
@@ -315,7 +315,7 @@ const ManuscriptSidePanel = (props: ManuscriptSidePanelProps) => {
               </>
             ) : null}
             {researchPanelTab === ResearchTabs.history ? <HistoryTab /> : null}
-            {researchPanelTab === ResearchTabs.source ? <SourceTab /> : null}
+            {researchPanelTab === ResearchTabs.credit ? <CreditTab /> : null}
             {isAdmin ? (
               <>
                 <div
